[OMPIRBuilder][debug] Fix debug info for variables in target region.
When a new function is created to handle OpenMP target region, the
variables used in it are passed as arguments. The scope and the location
of these variable needs to be updated according to the new function.
This PR makes the required changes and also add the required debug data
for the extra parameter that is passed for device case.