[OMPIRBuilder] Don't discard the debug record from entry block.
When we get a function back from CodeExtractor, we disard its entry block after coping its instructions into the entry block we prepared. While copying the instructions, the terminator is discarded for obvious reasons. But if there were some debug values attached to the terminator, those are useful and needs to be copied.

https://github.com/llvm/llvm-project/pull/135161